isomaker: Some fixes for bugs found when testing different buffer read sizes and read ordering. [build fix]
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@67862 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Runtime/common/fs/isomakercmd.cpp
@@ -776,7 +776,7 @@ static int rtFsIsoMakerCmdVerifyImageInRandomOrder(PRTFSISOMAKERCMDOPTS pOpts, R
              */
             uint32_t iBlock = RTRandU32Ex(0, cBlocks - 1);
             if (!ASMBitTestAndSet(pvBitmap, iBlock))
-                Assert(iBlock < (int32_t)cBlocks);
+                Assert(iBlock < cBlocks);
             else
             {
                 /* try 32 other random numbers. */
